+++ apps/meteor/client/sidebar/header/actions/CreateRoomList.tsx
@@ -1,14 +1,13 @@
 import { OptionTitle } from '@rocket.chat/fuselage';
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
-import { useSetModal, useSetting, useAtLeastOnePermission, useTranslation } from '@rocket.chat/ui-contexts';
-import React, { FC } from 'react';
+import { useSetting, useAtLeastOnePermission, useTranslation } from '@rocket.chat/ui-contexts';
+import React, { ReactElement, MouseEvent } from 'react';
 
-import { popover } from '../../../../app/ui-utils/client';
 import CreateDiscussion from '../../../components/CreateDiscussion';
 import ListItem from '../../../components/Sidebar/ListItem';
-import CreateTeamModal from '../../../views/teams/CreateTeamModal';
 import CreateChannelWithData from '../CreateChannelWithData';
 import CreateDirectMessage from '../CreateDirectMessage';
+import CreateTeam from '../CreateTeam';
+import { useCreateRoomModal } from '../hooks/useCreateRoomModal';
 
 const CREATE_CHANNEL_PERMISSIONS = ['create-c', 'create-p'];
 const CREATE_TEAM_PERMISSIONS = ['create-team'];
@@ -19,36 +18,19 @@ type CreateRoomListProps = {
 	closeList: () => void;
 };
 
-const useReactModal = (Component: FC<any>): ((e: React.MouseEvent<HTMLElement>) => void) => {
-	const setModal = useSetModal();
-
-	return useMutableCallback((e) => {
-		popover.close();
-
-		e.preventDefault();
-
-		const handleClose = (): void => {
-			setModal(null);
-		};
-
-		setModal(() => <Component onClose={handleClose} />);
-	});
-};
-
-const CreateRoomList: FC<CreateRoomListProps> = ({ closeList }) => {
+const CreateRoomList = ({ closeList }: CreateRoomListProps): ReactElement => {
 	const t = useTranslation();
+	const discussionEnabled = useSetting('Discussion_enabled');
 
 	const canCreateChannel = useAtLeastOnePermission(CREATE_CHANNEL_PERMISSIONS);
 	const canCreateTeam = useAtLeastOnePermission(CREATE_TEAM_PERMISSIONS);
 	const canCreateDirectMessages = useAtLeastOnePermission(CREATE_DIRECT_PERMISSIONS);
 	const canCreateDiscussion = useAtLeastOnePermission(CREATE_DISCUSSION_PERMISSIONS);
 
-	const createChannel = useReactModal(CreateChannelWithData);
-	const createTeam = useReactModal(CreateTeamModal);
-	const createDiscussion = useReactModal(CreateDiscussion);
-	const createDirectMessage = useReactModal(CreateDirectMessage);
-
-	const discussionEnabled = useSetting('Discussion_enabled');
+	const createChannel = useCreateRoomModal(CreateChannelWithData);
+	const createTeam = useCreateRoomModal(CreateTeam);
+	const createDiscussion = useCreateRoomModal(CreateDiscussion);
+	const createDirectMessage = useCreateRoomModal(CreateDirectMessage);
 
 	return (
 		<>
@@ -58,7 +40,7 @@ const CreateRoomList: FC<CreateRoomListProps> = ({ closeList }) => {
 					<ListItem
 						icon='hashtag'
 						text={t('Channel')}
-						action={(e: React.MouseEvent<HTMLElement>): void => {
+						action={(e: MouseEvent<HTMLElement>): void => {
 							createChannel(e);
 							closeList();
 						}}
@@ -68,7 +50,7 @@ const CreateRoomList: FC<CreateRoomListProps> = ({ closeList }) => {
 					<ListItem
 						icon='team'
 						text={t('Team')}
-						action={(e: React.MouseEvent<HTMLElement>): void => {
+						action={(e: MouseEvent<HTMLElement>): void => {
 							createTeam(e);
 							closeList();
 						}}
@@ -78,7 +60,7 @@ const CreateRoomList: FC<CreateRoomListProps> = ({ closeList }) => {
 					<ListItem
 						icon='balloon'
 						text={t('Direct_Messages')}
-						action={(e: React.MouseEvent<HTMLElement>): void => {
+						action={(e: MouseEvent<HTMLElement>): void => {
 							createDirectMessage(e);
 							closeList();
 						}}
@@ -88,7 +70,7 @@ const CreateRoomList: FC<CreateRoomListProps> = ({ closeList }) => {
 					<ListItem
 						icon='discussion'
 						text={t('Discussion')}
-						action={(e: React.MouseEvent<HTMLElement>): void => {
+						action={(e: MouseEvent<HTMLElement>): void => {
 							createDiscussion(e);
 							closeList();
 						}}
